fix: change Position to float

+++ src/Aevatar.Application.Contracts/Agents/WorkflowAgents.cs
@@ -12,6 +12,6 @@ public class WorkflowAgentDefinesDto
 {
     public string GrainId { get; set; }
     public string? NextGrainId { get; set; }
-    public int XPosition { get; set; }
-    public int YPosition { get; set; }
+    public float XPosition { get; set; }
+    public float YPosition { get; set; }
 }
